API/requirements_api.txt
pydantic
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
python-jose[cryptography]
argon2-cffi==23.1.0
pytest
testcontainers[postgres]
httpx
pandas
numpy
sentence_transformers
joblib

